% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.R
\name{acs_pct_transform}
\alias{acs_pct_transform}
\title{Calculate Proportions from Long-Formatted ACS Subsets by Population}
\usage{
acs_pct_transform(df, unique_col, percent = TRUE)
}
\arguments{
\item{df}{A dataframe.}

\item{unique_col}{Character or integer. Column containing unique values of
aggregating unit (i.e., the FIPS code).}

\item{percent}{If \code{TRUE} (default), calculates a percentage. Otherwise, returns a
proportion.}
}
\value{
A dataframe.
}
\description{
Calculate Proportions from Long-Formatted ACS Subsets by Population
}
