% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.R
\name{acs_pivot}
\alias{acs_pivot}
\title{Pivot ACS Estimates Long to Wide}
\usage{
acs_pivot(df, percent = TRUE)
}
\arguments{
\item{df}{A dataframe.}

\item{percent}{Whether to calculate percentages from long-formatted table.
Assumes that one value of \code{label} is a total.}
}
\value{
A dataframe.
}
\description{
Pivot ACS Estimates Long to Wide
}
