% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.R
\name{acs_process_nested}
\alias{acs_process_nested}
\title{Process Nested Data Structures in ACS Estimates.}
\usage{
acs_process_nested(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
A dataframe.
}
\description{
Process Nested Data Structures in ACS Estimates.
}
