% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_conn}
\alias{db_conn}
\title{Create connection to DB}
\usage{
db_conn(dbname, role, pass, host = "localhost", port = 5432)
}
\arguments{
\item{dbname}{Character. Name of database.}

\item{role}{Character. Name of role.}

\item{pass}{Character. Password.}

\item{host}{Character. Host address. Defaults to "localhost".}

\item{port}{Integer. Port number. Defaults to 5432.}
}
\value{
conn
}
\description{
Thin wrapper around RPostgres::dbConnect().
}
