% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_create}
\alias{db_create}
\title{Create Database}
\usage{
db_create(conn, dbname)
}
\arguments{
\item{conn}{Database connection like that created by \code{RPostgres::dbConnect()}.}

\item{dbname}{Character. Name of database.}
}
\value{
conn
}
\description{
Create Database
}
