% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_create_extension}
\alias{db_create_extension}
\title{Create Extension}
\usage{
db_create_extension(conn, ext)
}
\arguments{
\item{conn}{Database connection like that created by `RPostgres::dbConnect()`.}

\item{ext}{Character. Name of extension, e.g., "postgis."}
}
\value{
conn
}
\description{
Create Extension
}
