% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_create_extensions}
\alias{db_create_extensions}
\title{Create PostGIS extension.}
\usage{
db_create_extensions(conn, extensions = c("postgis", "postgis_raster"))
}
\arguments{
\item{conn}{Database connection like that created by `RPostgres::dbConnect()`.}

\item{extension}{Character. Name of extension to create.}
}
\value{
conn
}
\description{
Create PostGIS extension.
}
