% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_create_flow}
\alias{db_create_flow}
\title{Create DB and User, Set Access}
\usage{
db_create_flow(
  conn,
  dbname,
  role,
  role_pass,
  admin_pass,
  extensions = c("postgis", "postgis_raster")
)
}
\arguments{
\item{conn}{Database connection like that created by `RPostgres::dbConnect()`.}

\item{dbname}{Character. Name of database.}

\item{role}{Character. Name of role.}
}
\value{
conn
}
\description{
Create DB and User, Set Access
}
