% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_create_if}
\alias{db_create_if}
\title{Create DB and User, Set Access}
\usage{
db_create_if(conn, dbname, role, pass)
}
\arguments{
\item{conn}{Database connection like that created by `RPostgres::dbConnect()`.}

\item{dbname}{Character. Name of database.}

\item{role}{Character. Name of role.}
}
\description{
Create DB and User, Set Access
}
