% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_create_postgis}
\alias{db_create_postgis}
\title{Create PostGIS extension.}
\usage{
db_create_postgis(conn, dbname)
}
\arguments{
\item{conn}{Database connection like that created by `RPostgres::dbConnect()`.}

\item{dbname}{Character. Name of database.}
}
\description{
Create PostGIS extension.
}
