% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_create_postgis_raster}
\alias{db_create_postgis_raster}
\title{Create PostGIS Raster extension.}
\usage{
db_create_postgis_raster(conn, dbname)
}
\arguments{
\item{conn}{Database connection like that created by `RPostgres::dbConnect()`.}

\item{dbname}{Character. Name of database.}
}
\value{
conn
}
\description{
Create PostGIS Raster extension.
}
