% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_drop}
\alias{db_drop}
\title{Drop Database If It Exists}
\usage{
db_drop(conn, dbname)
}
\arguments{
\item{conn}{Database connection like that created by `RPostgres::dbConnect()`.}

\item{dbname}{Character. Name of database.}
}
\description{
Drop Database If It Exists
}
