% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_extension_exists}
\alias{db_extension_exists}
\title{Test Whether Extension Exists}
\usage{
db_extension_exists(conn, ext)
}
\arguments{
\item{conn}{Database connection like that created by `RPostgres::dbConnect()`.}

\item{ext}{Character. Name of extension, e.g., "postgis."}
}
\value{
Boolean.
}
\description{
Test Whether Extension Exists
}
