% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_grant_access}
\alias{db_grant_access}
\title{Grant DB Acess to Role}
\usage{
db_grant_access(conn, dbname, role)
}
\arguments{
\item{conn}{Database connection like that created by `RPostgres::dbConnect()`.}

\item{dbname}{Character. Name of database.}

\item{role}{Character. Name of role.}
}
\description{
Grant DB Acess to Role
}
