% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_role_create}
\alias{db_role_create}
\title{Create Role With a Password}
\usage{
db_role_create(conn, role, pass)
}
\arguments{
\item{conn}{Database connection like that created by \code{RPostgres::dbConnect()}.}

\item{role}{Character. Name of role.}

\item{pass}{Character. Password}
}
\value{
conn
}
\description{
Create Role With a Password
}
