% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_role_exists}
\alias{db_role_exists}
\title{Test Whether Role Exists}
\usage{
db_role_exists(conn, role)
}
\arguments{
\item{conn}{Database connection like that created by \code{RPostgres::dbConnect()}.}

\item{role}{Character. Name of role.}
}
\value{
Boolean
}
\description{
Test Whether Role Exists
}
