% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_set_defaults}
\alias{db_set_defaults}
\title{Set Default Access for Role on Database}
\usage{
db_set_defaults(conn, dbname, role)
}
\arguments{
\item{conn}{Database connection like that created by `RPostgres::dbConnect()`.}

\item{dbname}{Character. Name of database.}

\item{role}{Character. Name of role.}
}
\value{
conn
}
\description{
Set Default Access for Role on Database
}
