% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{detect_utm}
\alias{detect_utm}
\title{Detect UTM Zone}
\usage{
detect_utm(df)
}
\arguments{
\item{df}{Simple features data frame or tibble.}
}
\value{
Object of class `crs`.
}
\description{
Detect UTM Zone
}
