% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{elevate_points}
\alias{elevate_points}
\title{Add Z Dimension to Points from Column}
\usage{
elevate_points(df, zcol)
}
\arguments{
\item{df}{Simple features data frame or tibble.}

\item{zcol}{Character. Name of column containing Z values.}
}
\value{
Simple features data frame or tibble.
}
\description{
Add Z Dimension to Points from Column
}
