% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.R
\name{flow_terrain}
\alias{flow_terrain}
\title{Terrain Workflow}
\usage{
flow_terrain(
  extent,
  expand = NULL,
  interval = 10,
  z_scale = 3,
  tiles_on_side = 8,
  stepped = FALSE,
  mask = TRUE
)
}
\arguments{
\item{extent}{Simple features data frame or tibble (ideally polygon). Must be
projected and in linear units.}

\item{expand}{A numeric value of a distance, in map units, used to expand 
the bounding box that is used to fetch the terrain tiles.}

\item{interval}{Numeric. Elevation interval at which contours should be 
drawn. In units of crs.}

\item{z_scale}{How much to exaggerate the z scale. Default is 1. Values of 
3-4 are fairly standard in cartographic practice.}

\item{tiles_on_side}{Number of tiles on one side of extent. Must be the
result of 2 raised to a power or 1.}

\item{stepped}{Boolean. If `TRUE`, returns rasterized stepped topography 
based on contours.}
}
\value{
Named list
}
\description{
Terrain Workflow
}
