% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.R
\name{get_acs}
\alias{get_acs}
\alias{get_acs_industries}
\alias{get_acs_occupations}
\alias{get_acs_ancestry}
\alias{get_acs_place_of_birth}
\alias{get_acs_housing}
\alias{get_acs_race}
\alias{get_acs_age}
\title{Get ACS Variables by Census Unit}
\usage{
get_acs_industries(
  states,
  year,
  census_unit,
  county = NULL,
  crs = 4326,
  geometry = FALSE
)

get_acs_occupations(
  states,
  year,
  census_unit,
  county = NULL,
  crs = 4326,
  geometry = FALSE
)

get_acs_ancestry(
  states,
  year,
  census_unit,
  county = NULL,
  crs = 4326,
  geometry = FALSE
)

get_acs_place_of_birth(
  states,
  year,
  census_unit,
  county = NULL,
  crs = 4326,
  geometry = FALSE
)

get_acs_housing(states, year, census_unit, crs = 4326, geometry = FALSE)

get_acs_race(states, year, census_unit, crs = 4326, geometry = FALSE)

get_acs_age(states, year, census_unit, crs = 4326, geometry = FALSE)
}
\arguments{
\item{states}{A character vector of two-digit state abbreviations (e.g., `c("MA", "MI")).}

\item{year}{Integer. Final year of 5-year ACS period.}

\item{census_unit}{Character. The desired geography.}

\item{county}{County for which you are requesting data.}

\item{crs}{Target coordinate reference system: object of class `crs`, or input string for `st_crs`.}

\item{geometry}{If `FALSE` (the default), returns a table. If `TRUE`, returns a simple feature (`sf`) dataframe.}
}
\value{
A tibble or sf tibble of ACS data.
}
\description{
`get_acs_industries()` obtains data on employment by industry from ACS table [S2403 (Industry by Sex for the Civilian 
Employed Population 16 Years and Over)](https://data.census.gov/table/ACSST5Y2022.S2401).

`get_acs_occupations()` obtains data on civilian occupation from ACS table 
[S2401 (Occupation by Sex for the Civilan Employed Population 16 Years 
and Over)](https://data.census.gov/table/ACSST5Y2022.S2401).

`get_acs_ancestry()` obtains data on reported ancestry from ACS table 
[B04006 (People Reporting Ancestry)](https://data.census.gov/table/ACSDT5Y2023.B04006).

`get_acs_place_of_birth()` obtains data on place of birth for migrant population from ACS table
[B05006 (Place of Birth for the Foreign-Born Population in the United States)](https://data.census.gov/table/ACSDT5Y2023.B05006).

`get_acs_housing()`obtains data on tenure, housing type, and gross rent from ACS tables
[B25042 (Tenure by Bedrooms)](https://data.census.gov/table/ACSDT5Y2023.B25042), [B25024 (Units in Structure)](https://data.census.gov/table/ACSDT5Y2023.B25024), and
[B25031 (Median Gross Rent by Bedrooms)](https://data.census.gov/table/ACSDT5Y2023.B25031).

`get_acs_race()` Obtains data on race and ethnicity, including median household and per-capita income measures, from tables 
[B03002 (Hispanic or Latino Origin by Race)](https://data.census.gov/table/ACSDT5Y2023.B03002), [B19301B-H (Per capita income in the past 12 months)](https://data.census.gov/table/ACSDT5Y2023.B19301B), 
and [B19013B-H (Median household income in the past 12 months)](https://data.census.gov/table/ACSDT5Y2023.B19013B).

`get_acs_age()` obtains data on age, sliced by sex, from ACS table [B01001 (Sex by Age)](https://data.census.gov/table/ACSDT5Y2023.B01001).
}
