% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm.R
\name{get_osm}
\alias{get_osm}
\alias{osm_extent_to_opq}
\alias{osm_get_buildings}
\alias{osm_get_open_space}
\title{Get OpenStreetMap Features by Key}
\usage{
osm_extent_to_opq(extent)

osm_get_buildings(extent)

osm_get_open_space(extent)
}
\arguments{
\item{extent}{`sf` object used to generate `bbox`}

\item{key}{OpenStreetMap feature key.}
}
\value{
`sf` object
}
\description{
`get_osm_key()` retrieves an arbitrary feature type based on its key.
`get_osm_buildings()` retrieves buildings based on bounding box of a feature.
}
