% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs.R
\name{gtfs_feeds}
\alias{gtfs_feeds}
\alias{gtfs_get_mbta}
\title{Retrieve Various GTFS Feeds}
\usage{
gtfs_get_mbta(crs)
}
\arguments{
\item{crs}{`crs` object or EPSG code.}
}
\value{
A list including `routes` and `stops` as `sf` objects.
}
\description{
`gtfs_mbta()` retrieves the GTFS feed of the Massachusetts Bay Transportation
Authority (MBTA) servicing eastern Massachusetts.

`gtfs_mta()` retrieves the GTFS feed of the Metropolitan Transportation
Authority (MTA) servicing the New York City metro area.
}
