% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lodes.R
\name{lodes_get_data}
\alias{lodes_get_data}
\title{Download LODES OD tables by census unit.}
\usage{
lodes_get_data(states, year, census_unit)
}
\arguments{
\item{states}{Character vector of state abbreviations.}

\item{year}{Integer. Year of LODES data.}

\item{census_unit}{One of c("bg", "tract", "county", "state").}
}
\value{
A dataframe of LODES files.
}
\description{
Download LODES OD tables by census unit.
}
