% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munis.R
\name{munis_decision}
\alias{munis_decision}
\title{Execute Municipality-Getter Using State Name}
\usage{
munis_decision(states, crs)
}
\arguments{
\item{states}{Character vector of state names.}

\item{crs}{target coordinate reference system.}
}
\value{
An `sf` object
}
\description{
Execute Municipality-Getter Using State Name
}
