% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munis.R
\name{munis_get_*}
\alias{munis_get_*}
\alias{munis_get_ma}
\title{Downloads Municipal Boundaries by State}
\usage{
munis_get_ma()
}
\value{
An `sf` object.
}
\description{
`muni_get_ma()` Downloads municipal boundaries for Massachusetts.

`muni_get_ct()` Downloads municipal boundaries for Connecticut.

`muni_get_me()` Downloads municipal boundaries for Maine.

`muni_get_nh()` Downloads municipal boundaries for New Hampshire.

`muni_get_nj()` Downloads municipal boundaries for New Jersey.

`muni_get_ny()` Downloads municipal boundaries for New York.

`muni_get_ri()` Downloads municipal boundaries for Rhode Island.

`muni_get_vt()` Downloads municipal boundaries for Vermont.
}
