% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munis.R
\name{munis_get_munis}
\alias{munis_get_munis}
\title{Get Municipality by State and Munis}
\usage{
munis_get_munis(places, crs = 4326, filter = TRUE, fallbacks = c("cdp", "osm"))
}
\arguments{
\item{places}{Character vector of state names.}

\item{crs}{target coordinate reference system.}
}
\value{
An `sf` object
}
\description{
Get Municipality by State and Munis
}
