% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munis.R
\name{munis_preprocess}
\alias{munis_preprocess}
\title{Preprocessing Workflow for Municipalities}
\usage{
munis_preprocess(df, crs)
}
\arguments{
\item{df}{An `sf` object.}

\item{crs}{Target coordinate reference system.}
}
\value{
An `sf` object.
}
\description{
Preprocessing Workflow for Municipalities
}
