% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm.R
\name{osm_get_}
\alias{osm_get_}
\alias{osm_get_features}
\alias{osm_get_key}
\alias{osm_get_buildings}
\alias{osm_get_open_space}
\alias{osm_get_bike_lanes}
\title{Get OpenStreetMap Features}
\usage{
osm_get_features(extent, features, type)

osm_get_key(extent, key, type)

osm_get_buildings(extent)

osm_get_open_space(extent)

osm_get_bike_lanes(extent)
}
\arguments{
\item{extent}{\code{sf} object used to generate the overpass query.}

\item{features}{Named list or vector with key value pairs.}

\item{type}{String. "Polygons" or "lines".}

\item{key}{A feature key (e.g., "building").}
}
\value{
\code{sf} object
}
\description{
\code{osm_get_buildings()} Retrieves buildings from OpenStreetMap, given an
input extent.

\code{osm_get_open_space()} Retrieves open space from OpenStreetMap, given an
input extent.

\code{osm_get_bike_lanes()} Retrieves bike lanes from OpenStreetMap, given an
input extent.
}
