% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyplanr.R
\name{shinyplanr}
\alias{shinyplanr}
\title{ShinyPlanR}
\usage{
shinyplanr()
}
\value{
Nothing.
}
\description{
Run a shiny interface walking through configuration of a batch download
workflow.
}
\examples{
shinyplanr()
}
