% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_bbox_sf}
\alias{st_bbox_sf}
\title{Bounding Box as `sf` Object}
\usage{
st_bbox_sf(df)
}
\arguments{
\item{df}{An `sf` object.}
}
\value{
An `sf` object.
}
\description{
Given an input `sf` object, returns that object's bounding box as an `sf`
object.
}
