% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_clip}
\alias{st_clip}
\title{GIS-esque Clip Function}
\usage{
st_clip(x, y)
}
\arguments{
\item{x}{Layer to be clipped.}

\item{y}{Layer to do the clipping.}
}
\value{
Layer \code{x} clipped to \code{y}.
}
\description{
GIS-esque Clip Function
}
