% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_contours}
\alias{st_contours}
\title{Contours from Raster}
\usage{
st_contours(
  raster,
  interval,
  maxcells = terra::ncell(raster)/4,
  threshold_length = units::as_units(250, "m")
)
}
\arguments{
\item{raster}{\code{RasterLayer}, for example returned by \code{st_get_dem()}.}

\item{interval}{Numeric. Elevation interval at which contours should be
drawn. In units of crs.}

\item{maxcells}{Maximum number of raster cells to use. Reduce if function is
failing.}

\item{threshold_length}{Threshold length below which contours will be
dropped. Helps to remove shards. Defaults to \verb{units::as_units(250, "m").}}
}
\value{
\code{sf} object with \code{LINESTRING} geometries.
}
\description{
Generally used to create contour lines from a digital elevation model (DEM).
}
