% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_contours_enclose}
\alias{st_contours_enclose}
\title{Enclose Contours Using Polygon Features}
\usage{
st_contours_enclose(contours, enclosure_layer, poly = FALSE, bbox = FALSE)
}
\arguments{
\item{contours}{`sf` object, as returned by `st_contour()`.}

\item{enclosure_layer}{Feature layer used to enclose contour lines.}

\item{poly}{Boolean. If `FALSE` (default), returns `MULTILINESTRING`s. If 
`TRUE`, returns `MULTIPOLYGON`s.}

\item{bbox}{Boolean. If `FALSE` (default), contours are enclosed by feature
boundaries. If `TRUE`, contours are enclosed by their bounding box.}
}
\value{
`sf` object, `POLYGON` geometries.
}
\description{
Enclose Contours Using Polygon Features
}
