% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_contours_model}
\alias{st_contours_model}
\title{Elevate Contours and Finish}
\usage{
st_contours_model(enclosed_contours, full_plates = TRUE, etch_guides = FALSE)
}
\arguments{
\item{enclosed_contours}{`sf` object, as returned by `st_contours_close()`}

\item{full_plates}{If `TRUE` (default), contours extended to the edge of the
model. If `FALSE`, they will emeerge in 'strips'.}

\item{etch_guides}{If `TRUE`, guides are created by etching edge of the above
level into each level. If `FALSE` (default), no guides created. Requires that
`poly` and `full_plates` both be `TRUE`.}
}
\value{
`sf` object.
}
\description{
This function is primarily used in the case where the user is preparing a
stepped contour model in CAD (or using a laser cutter).
}
