% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_detect_utm}
\alias{st_detect_utm}
\title{Detect UTM Zone Based on Location}
\usage{
st_detect_utm(df)
}
\arguments{
\item{df}{Simple features data frame or tibble.}
}
\description{
Detect UTM Zone Based on Location
}
