% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_filter_and_join}
\alias{st_filter_and_join}
\title{Runs a series of standard subsetting steps}
\usage{
st_filter_and_join(
  df,
  filter = NULL,
  clip = NULL,
  join = NULL,
  int_join = FALSE
)
}
\arguments{
\item{df}{`sf` object}

\item{filter}{`sf` object to spatially filter by.}

\item{clip}{`sf` object to clip input by.}

\item{join}{`sf` object to spatially join by.}

\item{int_join}{If FALSE (default), `join` is a left join. If `int_join`, 
it is intersected with the input.}
}
\value{
A processed `sf` object.
}
\description{
Runs a series of standard subsetting steps
}
