% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_get_dem}
\alias{st_get_dem}
\title{Get DEM from AWS Terrain Tiles}
\usage{
st_get_dem(
  extent,
  tiles_on_side = 4,
  expand = NULL,
  z = st_zoom_from_extent(extent = extent, tiles_on_side = tiles_on_side),
  src = "aws"
)
}
\arguments{
\item{extent}{Simple features data frame or tibble (ideally polygon). Must be
projected and in linear units.}

\item{tiles_on_side}{Number of tiles on one side of extent. Must be the
result of 2 raised to a power or 1.}

\item{expand}{A numeric value of a distance, in map units, used to expand 
the bounding box that is used to fetch the terrain tiles.}

\item{z}{Number. Zoom level. If specified, `tiles_on_side` is ignored. If not
specified, the result of `st_zoom_from_extent(extent, tiles_on_side)`. Note
that `elevatr::get_elev_raster()` has a max `z` of 14.}

\item{src}{One of "aws", "gl3", "gl1", "alos", "srtm15plus".}
}
\value{
A `RasterLayer`.
}
\description{
Get DEM from AWS Terrain Tiles
}
