% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_is_sf}
\alias{st_is_sf}
\title{`stop()` if  Object is Not `sf`}
\usage{
st_is_sf(df)
}
\arguments{
\item{df}{Dataframe or `sf` object.}
}
\value{
Nothing.
}
\description{
`stop()` if  Object is Not `sf`
}
