% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_is_type}
\alias{st_is_type}
\alias{st_is_multipolygon}
\alias{st_is_polygon}
\alias{st_is_multilinestring}
\alias{st_is_linestring}
\alias{st_is_multipoint}
\alias{st_is_point}
\title{Tests \code{sf} Dataframes for Geometry Types}
\usage{
st_is_type(df, type, exact = FALSE, by_geometry = FALSE, ...)

st_is_multipolygon(df, by_geometry = FALSE, ...)

st_is_polygon(df, by_geometry = FALSE, ...)

st_is_multilinestring(df, by_geometry = FALSE, ...)

st_is_linestring(df, by_geometry = FALSE, ...)

st_is_multipoint(df, by_geometry = FALSE, ...)

st_is_point(df, by_geometry = FALSE, ...)
}
\arguments{
\item{df}{An \code{sf} dataframe.}

\item{type}{Character. Type to match.}

\item{exact}{Boolean. If \code{FALSE} (default), returns unexact matches. (E.g.,
\code{"POLYGON"} will also match \verb{"MULTIPOLYGON".})}

\item{by_geometry}{If \code{FALSE} (default), returns a single result for the
entire dataframe. (I.e., will be a "GEOMETRYCOLLECTION" if types are mixed).
Otherwise, returns a rowwise result.}

\item{...}{Additional arguments passed to \code{st_is_type()}.}
}
\value{
Boolean.
}
\description{
\code{st_is_type()} returns \code{TRUE} if \code{type} matches geometry type returned by
\code{sf::st_geometry_type()}.

\code{st_is_multipolygon()} returns \code{TRUE} if geometry column contains
MULTIPOLYGONS.

\code{st_is_polygon()} returns \code{TRUE} if geometry column contains
POLYGONS (or MULTIPOLYGONS, if \code{exact} is FALSE).

\code{st_is_multilinestring()} returns \code{TRUE} if geometry column contains
MULTILINESTRINGS.

\code{st_is_linestring()} returns \code{TRUE} if geometry column contains
LINESTRINGS (or MULTILINESTRINGS if \code{exact} is FALSE).

\code{st_is_multipoint()}  returns \code{TRUE} if geometry column contains
MULTIPOINTS.

\code{st_is_point()} returns \code{TRUE} if geometry column contains
POINTS (or MULTIPOINTS if \code{exact} is FALSE).
}
