% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_make_line}
\alias{st_make_line}
\title{Make Line from Coordinate Pair}
\usage{
st_make_line(xyxy)
}
\arguments{
\item{xyxy}{Vector of coordinate pair in the form c(x1, y1, x2, y2)}
}
\value{
A LINESTRING.
}
\description{
Make Line from Coordinate Pair
}
