% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_model_scale}
\alias{st_model_scale}
\title{Scale Layer Based on Scaling Parameters}
\usage{
st_model_scale(df, scale, thickness, z_scale, contour_int)
}
\arguments{
\item{df}{Simple features dataframe or tibble.}

\item{scale}{Scaling parameters, as returend by `scale_to_model()`.}

\item{thickness}{Numeric. Thickness of model plywood sheets.}

\item{z_scale}{Numeric. Z-exaggeration for elevation.}

\item{contour_int}{Nuemeric. Intervals for contours.}
}
\value{
Scaled simple features dataframe or tibble.
}
\description{
Scale Layer Based on Scaling Parameters
}
