% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_multi_type_center}
\alias{st_multi_type_center}
\title{Return Geometry Centers for Multiple Types.}
\usage{
st_multi_type_center(df, on_surface = TRUE, retain_geom = FALSE)
}
\arguments{
\item{df}{A dataframe with point geometries.}

\item{on_surface}{If `TRUE` (the default), generates polygon centroids using 
`st_point_on_surface()`. If `FALSE`, uses `st_centroid()`.}

\item{retain_geom}{Boolean. If `TRUE` (the default), returns points. If 
`FALSE` preserves geometry type of input.}
}
\value{
An `sf` dataframe.
}
\description{
Return Geometry Centers for Multiple Types.
}
