% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_point_from_coords}
\alias{st_point_from_coords}
\title{Create a Point Dataframe Row from Coordinate Pair}
\usage{
st_point_from_coords(coords, crs, name = NULL, coord_crs = 4326)
}
\arguments{
\item{coords}{vector of longitude, latitude coordinates (in EPSG:4326, unless
coord_crs is changed).}

\item{crs}{EPSG code or \code{crs} object.}

\item{name}{Optional. Name to give point (in name field).}

\item{coord_crs}{\code{crs} of provided points. 4326 default.}
}
\value{
\code{sf} object
}
\description{
Create a Point Dataframe Row from Coordinate Pair
}
