% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_regionalize}
\alias{st_regionalize}
\title{Build Regional Polygon from Input Layer}
\usage{
st_regionalize(df, dist, type)
}
\arguments{
\item{df}{\code{sf} object}

\item{dist}{\code{units} object or number, in which case output will be based on
CRS of input.}

\item{type}{One of "bbox", "square", or "circle".}
}
\value{
\code{sf} object
}
\description{
Builds a "region" on the basis of an input layer's extent, a buffer distance,
and a geometric type.
}
