% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_scale_to_model}
\alias{st_scale_to_model}
\title{Retrieve Scaling Parameters, Given Model Size}
\usage{
st_scale_to_model(df, model_size = NULL)
}
\arguments{
\item{df}{\code{sf} object}

\item{model_size}{Optional. Largest dimension of model, as numeric or
\code{units}. If numeric, treated as map units. If not provided, model will not be
scaled.}
}
\value{
Named list with \code{factor}, \code{x_delta}, \code{y_delta}.
}
\description{
Retrieve Scaling Parameters, Given Model Size
}
