% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_simple}
\alias{st_simple}
\alias{st_width}
\alias{st_height}
\alias{st_middle}
\alias{st_max_dim}
\title{Calculate Simple Measurements of \code{sf} Objects}
\usage{
st_width(df)

st_height(df)

st_middle(df)

st_max_dim(df)
}
\arguments{
\item{df}{\code{sf} object}
}
\value{
A number for \code{st_width} and \code{st_height}. A named list for
\code{st_middle}.
}
\description{
\code{st_width()} returns the bounding width in units of the crs.
\code{st_height()} returns the bounding height in units of the crs.
\code{st_middle()} returns the center of the bounding box.
\code{st_max_dim()} returns the largest dimension, named 'x' or 'y'.
}
