% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_square_it}
\alias{st_square_it}
\title{Bounding Square, Bounding Circle}
\usage{
st_square_it(df, circle = FALSE)
}
\arguments{
\item{df}{\code{sf} object}

\item{circle}{If \code{TRUE}, returns a circle. If \code{FALSE} (the default), returns
a square.}
}
\value{
A circle or square \code{sf} object whose size is given by the larger
dimension of a layer's bounding box.
}
\description{
Bounding Square, Bounding Circle
}
