% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_upgrade_type}
\alias{st_upgrade_type}
\title{Upgrade Geometry Type}
\usage{
st_upgrade_type(df)
}
\arguments{
\item{df}{`sf` object}
}
\value{
An `sf` object, with upcast geometry.
}
\description{
Given an input `sf` object that is made up "POINT"s, "POLYLINE"s, etc.,
upcasts that object to "MULTIPOINT", "MULTIPOLYLINE", etc.
}
