% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_zoom_from_extent}
\alias{st_zoom_from_extent}
\title{Determine Zoom Level from Extent and Requested Tile Resolution}
\usage{
st_zoom_from_extent(df, tiles_on_side = 2)
}
\arguments{
\item{df}{`sf` object}

\item{tiles_on_side}{Numeric. Tile 'resolution' (i.e., number of map tiles on
the longest side of the extent's bounding box). Must be two raised to a power
or 1.}
}
\value{
An integer zoom level.
}
\description{
Determine Zoom Level from Extent and Requested Tile Resolution
}
