% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigris.R
\name{tigris_get_}
\alias{tigris_get_}
\alias{tigris_get_multi}
\alias{tigris_get_states}
\alias{tigris_get_counties}
\alias{tigris_get_places}
\alias{tigris_get_postal}
\alias{tigris_get_tracts}
\alias{tigris_get_block_groups}
\alias{tigris_get_roads}
\alias{tigris_get_rails}
\alias{tigris_get_area_water}
\alias{tigris_get_linear_water}
\title{Get Census Geographies}
\usage{
tigris_get_multi(
  .function,
  places,
  state = NULL,
  county = NULL,
  year = NULL,
  crs = 4326,
  ...
)

tigris_get_states(places, year = NULL, crs = 4326, filter = TRUE)

tigris_get_counties(places, year = NULL, crs = 4326, filter = TRUE)

tigris_get_places(places, year = NULL, crs = 4326, filter = TRUE)

tigris_get_postal(places = NULL, year = 2010, crs = 4326, counties = NULL)

tigris_get_tracts(places, year = NULL, crs = 4326)

tigris_get_block_groups(places, year = NULL, crs = 4326)

tigris_get_roads(places, year = NULL, crs = 4326, counties = NULL)

tigris_get_rails(places, year = NULL, crs = 4326)

tigris_get_area_water(places, year = NULL, crs = 4326)

tigris_get_linear_water(places, year = NULL, crs = 4326)
}
\arguments{
\item{.function}{`tigris` function to run.}

\item{crs}{EPSG code or `crs` object. `4326` default.}

\item{...}{Passed on to `tigris` download function (e.g.,
`tigris::tracts()`).}

\item{counties}{Character vector of county names. If
`NULL` (the default), all counties are returned.}

\item{states}{Vector of two-character state codes.}

\item{starts_with}{Used only in `tigris_get_zcta()`. Beginning digits of
ZCTAs you want to return.}
}
\value{
Simple Features dataframe.
}
\description{
`tigris_get_places()` retrieves census-designated places.

`tigris_get_states()` retrieves states, commonwealths, and territories.

`tigris_get_counties()` retrieves counties and equivalents.

`tigris_get_zctas()` retrieves ZIP code tabulation areas.

`tigris_get_multistate()` is a helper function that downloads multiple states
worth of data, where `tigris` function prevents this.

`tigris_get_multistate_by_county()` is a helper function that downloads
multiple states worth of data, county-by-county, where `tigris` function
prevents this.

`tigris_get_tracts()` retrieves census tracts.

`tigris_get_block_groups()` retrieves block_groups.

`tigris_get_roads()` retrieves roads.

`tigris_get_primary_roads()` retrieves primary roads.

`tigris_get_primary_secondary_roads()` retrieves primary/secondary roads.

`tigris_get_area_water()` retrieves areal water features.

`tigris_get_linear_water()` retrieves linear water features.
}
