% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigris.R
\name{tigris_preprocess}
\alias{tigris_preprocess}
\title{Pre-processing operations for spatial data.}
\usage{
tigris_preprocess(df, crs)
}
\arguments{
\item{df}{Simple features dataframe.}

\item{crs}{EPSG code or `crs` object.}
}
\value{
Simple Features dataframe.
}
\description{
Pre-processing operations for spatial data.
}
