% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils_extent_to_census}
\alias{utils_extent_to_census}
\title{Extent to State/County}
\usage{
utils_extent_to_census(extent)
}
\arguments{
\item{extent}{\code{sf} object.}
}
\value{
A named list of counties that the.extent intersects. Takes the form
\code{list("MI" = c("Saginaw"))}.
}
\description{
Extent to State/County
}
