% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils_get_arc}
\alias{utils_get_arc}
\title{Read Remote ArcGIS Open Data Layer}
\usage{
utils_get_arc(id)
}
\arguments{
\item{id}{Character. ArcGIS online ID.}
}
\value{
Object of class `sf`.
}
\description{
Read Remote ArcGIS Open Data Layer
}
