% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils_get_remote}
\alias{utils_get_remote}
\title{Get Remote File and Write to Disk}
\usage{
utils_get_remote(url, path)
}
\arguments{
\item{url}{URL of remote file.}

\item{path}{Path to saved file.}
}
\value{
A \code{response()} object.
}
\description{
Get Remote File and Write to Disk
}
