% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{utils_get_remote_shp}
\alias{utils_get_remote_shp}
\title{Read Remote Shapefile Stored in `.zip` File}
\usage{
utils_get_remote_shp(url, layer)
}
\arguments{
\item{url}{URL of remote file containing shapefiles.}

\item{layer}{Name of Shapefile to read.}
}
\value{
Object of class `sf`.
}
\description{
Read Remote Shapefile Stored in `.zip` File
}
