% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{utils_get_zipped_shp}
\alias{utils_get_zipped_shp}
\title{Read Single Shapefile from `.zip` File}
\usage{
utils_get_zipped_shp(path, layer)
}
\arguments{
\item{path}{Path to `.zip` file.}

\item{layer}{Name of shapefile to read.}
}
\value{
Object of class `sf`.
}
\description{
Given a `.zip` file that conceivably contains multiple shapefiles, read only
one of them.
}
