% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{utils_prompt_check}
\alias{utils_prompt_check}
\title{Boolean Checking Prompt}
\usage{
utils_prompt_check(prompt)
}
\arguments{
\item{prompt}{Text of user prompt.}
}
\value{
Boolean user response.
}
\description{
Little helper that prompts a user and accepts binary responses.
}
