% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{utils_std_output_format}
\alias{utils_std_output_format}
\title{Standardize Output File Format}
\usage{
utils_std_output_format(format)
}
\arguments{
\item{format}{A common name of shapefiles, geopackages, geojsons, or postgis.}
}
\value{
Standardized character.
}
\description{
Standardize Output File Format
}
