% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils_write_multi}
\alias{utils_write_multi}
\title{Write Dataframe to One or Multiple Common Output Formats}
\usage{
utils_write_multi(data, name, dir_db, format, conn = NULL)
}
\arguments{
\item{data}{Dataframe or data frame extension (e.g., tibble).}

\item{name}{Character. Name of output table or file.}

\item{dir_db}{Character. Name of directory or database.}

\item{format}{Character vector including one or multiple of \code{"gpkg"},
\code{"postgis"}, \code{"dxf"}, or \code{"csv"}.}

\item{conn}{Database connection like that created by \code{RPostgres::dbConnect()}.}
}
\value{
Original dataframe.
}
\description{
Write Dataframe to One or Multiple Common Output Formats
}
