% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lodes.R
\name{xyxy_to_lines}
\alias{xyxy_to_lines}
\title{Takes a dataframe with four columns containing two XY pairs (e.g., [X_{1}, 
Y_{1}], [X_{2}, Y{2}]) and returns an sf dataframe with the same number of
rows with those coordinates converted into linestrings.}
\usage{
xyxy_to_lines(df, crs, names = c("x_h", "y_h", "x_w", "y_w"))
}
\arguments{
\item{df}{A dataframe with four columns containing two XY pairs.}

\item{crs}{Coordinate reference system.}

\item{names}{list of the names of the four columns.}
}
\value{
A dataframe with LINESTRING geometries.
}
\description{
Takes a dataframe with four columns containing two XY pairs (e.g., [X_{1}, 
Y_{1}], [X_{2}, Y{2}]) and returns an sf dataframe with the same number of
rows with those coordinates converted into linestrings.
}
