% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.R
\name{acs_get_vars}
\alias{acs_get_vars}
\title{Get 5-Year ACS Estimates Based on Variable List}
\usage{
acs_get_vars(
  vars,
  states,
  year,
  census_unit,
  crs = 4326,
  county = NULL,
  geometry = TRUE,
  drop_moe = TRUE
)
}
\arguments{
\item{vars}{Character vector of variables (see \code{tidycensus::load_variables}).}

\item{states}{Character vector of state abbreviations.}

\item{year}{Integer. Final year of ACS estimates.}

\item{census_unit}{Census unit for which to download tables.}

\item{crs}{Target coordinate reference system: object of class \code{crs}.}

\item{county}{Character. County name.}

\item{geometry}{Boolean. If \code{TRUE} (default) retains Tiger/LINE geography.}

\item{drop_moe}{Boolean. If \code{TRUE} (default) drops margin of error estimates.}
}
\value{
\code{sf} object if \code{geometry} is \code{TRUE}. Else data frame.
}
\description{
Get 5-Year ACS Estimates Based on Variable List
}
