% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_create_extension}
\alias{db_create_extension}
\title{Create PostGIS extension.}
\usage{
db_create_extension(conn, dbname, extension)
}
\arguments{
\item{conn}{Database connection like that created by `RPostgres::dbConnect()`.}

\item{dbname}{Character. Name of database.}

\item{extension}{Character. Name of extension to create.}
}
\value{
conn
}
\description{
Create PostGIS extension.
}
