% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_create_extensions}
\alias{db_create_extensions}
\title{Create extensions.}
\usage{
db_create_extensions(conn, exts = c("postgis", "postgis_raster"))
}
\arguments{
\item{conn}{Database connection like that created by \code{RPostgres::dbConnect()}.}

\item{exts}{Character vector. Name of extensions to create.}
}
\value{
conn
}
\description{
Create extensions.
}
