% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm.R
\name{get_osm}
\alias{get_osm}
\alias{osm_get_key}
\alias{osm_get_buildings}
\title{Get OpenStreetMap Features by Key}
\usage{
osm_get_key(extent, key)

osm_get_buildings(extent)
}
\arguments{
\item{extent}{`sf` object used to generate `bbox`}

\item{key}{OpenStreetMap feature key.}
}
\value{
`sf` object
}
\description{
`get_osm_key()` retrieves an arbitrary feature type based on its key.
`get_osm_buildings()` retrieves buildings based on bounding box of a feature.
}
