% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs.R
\name{gtfs_feeds}
\alias{gtfs_feeds}
\alias{gtfs_get_mbta}
\alias{gtfs_get_mta}
\title{Retrieve Various GTFS Feeds}
\usage{
gtfs_get_mbta(crs)

gtfs_get_mta(crs)
}
\arguments{
\item{crs}{\code{crs} object or EPSG code.}
}
\value{
A list including \code{routes} and \code{stops} as \code{sf} objects.
}
\description{
\code{gtfs_mbta()} retrieves the GTFS feed of the Massachusetts Bay Transportation
Authority (MBTA) servicing eastern Massachusetts.

\code{gtfs_mta()} retrieves the GTFS feed of the Metropolitan Transportation
Authority (MTA) servicing the New York City metro area.
}
