% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs.R
\name{gtfs_helpers}
\alias{gtfs_helpers}
\alias{gtfs_routes}
\alias{gtfs_stops}
\alias{gtfs_process}
\title{GTFS Helper Functions}
\usage{
gtfs_routes(gtfs, crs)

gtfs_stops(gtfs, crs)

gtfs_process(gtfs, crs)
}
\arguments{
\item{gtfs}{A `tidygtfs` object.}

\item{crs}{`crs` object or EPSG code.}
}
\value{
An `sf` object
}
\description{
`gtfs_routes()` takes a `tidygtfs` object and returns routes as `sf`.
`gtfs_stops()` takes a `tidygtfs` object and returns stops as `sf`.
`gtfs_process()` removes stops with no lat-longs and builds `sf` objects.
}
