% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munis.R
\name{muni_get_state}
\alias{muni_get_state}
\alias{munis_get_ma}
\alias{munis_get_ct}
\alias{munis_get_me}
\alias{munis_get_nh}
\alias{munis_get_nj}
\alias{munis_get_ny}
\alias{munis_get_ri}
\alias{munis_get_vt}
\title{Downloads Municipal Boundaries by State}
\usage{
munis_get_ma()

munis_get_ct()

munis_get_me()

munis_get_nh()

munis_get_nj()

munis_get_ny()

munis_get_ri()

munis_get_vt()
}
\value{
An `sf` object.
}
\description{
`muni_get_ma()` Downloads municipal boundaries for Massachusetts.

`muni_get_ct()` Downloads municipal boundaries for Connecticut.

`muni_get_me()` Downloads municipal boundaries for Maine.

`muni_get_nh()` Downloads municipal boundaries for New Hampshire.

`muni_get_nj()` Downloads municipal boundaries for New Jersey.

`muni_get_ny()` Downloads municipal boundaries for New York.

`muni_get_ri()` Downloads municipal boundaries for Rhode Island.

`muni_get_vt()` Downloads municipal boundaries for Vermont.
}
