% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_check_for_proj}
\alias{st_check_for_proj}
\title{Check if Layer Projection Matches CRS and Transform if Not}
\usage{
st_check_for_proj(df, crs = 4326)
}
\arguments{
\item{df}{`sf` object}

\item{crs}{EPSG code or `crs` object.}
}
\value{
Reprojected `sf` object.
}
\description{
Check if Layer Projection Matches CRS and Transform if Not
}
