% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_contours_enclose}
\alias{st_contours_enclose}
\title{Enclose Contours Using Polygon Features}
\usage{
st_contours_enclose(contours, enclosure_layer, poly = FALSE, bbox = FALSE)
}
\arguments{
\item{contours}{\code{sf} object, as returned by \code{st_contour()}.}

\item{enclosure_layer}{Feature layer used to enclose contour lines.}

\item{poly}{Boolean. If \code{FALSE} (default), returns \code{MULTILINESTRING}s. If
\code{TRUE}, returns \code{MULTIPOLYGON}s.}

\item{bbox}{Boolean. If \code{FALSE} (default), contours are enclosed by feature
boundaries. If \code{TRUE}, contours are enclosed by their bounding box.}
}
\value{
\code{sf} object, \code{POLYGON} geometries.
}
\description{
Enclose Contours Using Polygon Features
}
