% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_contours_model}
\alias{st_contours_model}
\title{Elevate Contours and Finish}
\usage{
st_contours_model(enclosed_contours, full_plates = TRUE, etch_guides = FALSE)
}
\arguments{
\item{enclosed_contours}{\code{sf} object, as returned by \code{st_contours_close()}}

\item{full_plates}{If \code{TRUE} (default), contours extended to the edge of the
model. If \code{FALSE}, they will emeerge in 'strips'.}

\item{etch_guides}{If \code{TRUE}, guides are created by etching edge of the above
level into each level. If \code{FALSE} (default), no guides created. Requires that
\code{poly} and \code{full_plates} both be \code{TRUE}.}
}
\value{
\code{sf} object.
}
\description{
This function is primarily used in the case where the user is preparing a
stepped contour model in CAD (or using a laser cutter).
}
