% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_filter_and_join}
\alias{st_filter_and_join}
\title{Runs a series of standard subsetting steps}
\usage{
st_filter_and_join(
  df,
  filter = NULL,
  clip = NULL,
  join = NULL,
  int_join = FALSE
)
}
\arguments{
\item{df}{\code{sf} object}

\item{filter}{\code{sf} object to spatially filter by.}

\item{clip}{\code{sf} object to clip input by.}

\item{join}{\code{sf} object to spatially join by.}

\item{int_join}{If FALSE (default), \code{join} is a left join. If \code{int_join},
it is intersected with the input.}
}
\value{
A processed \code{sf} object.
}
\description{
Runs a series of standard subsetting steps
}
