% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_get_dem}
\alias{st_get_dem}
\title{Get DEM from AWS Terrain Tiles}
\usage{
st_get_dem(area, z = 14, src = "aws")
}
\arguments{
\item{area}{Simple features data frame or tibble (ideally polygon).}

\item{z}{Zoom level to return.}

\item{src}{One of "aws", "gl3", "gl1", "alos", "srtm15plus".}
}
\value{
A `RasterLayer`.
}
\description{
Get DEM from AWS Terrain Tiles
}
