% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_hillshade}
\alias{st_hillshade}
\title{Calculate a Hillshade from a Digital Elevation Model.}
\usage{
st_hillshade(
  dem,
  angle = 45,
  direction = 300,
  normalize = TRUE,
  overwrite = TRUE,
  filename = ""
)
}
\arguments{
\item{dem}{A digital elevation model stored as a `RasterLayer`.}

\item{angle}{Elevation angle of the light source (degrees). Defaults to 45.}

\item{direction}{Azimuth angle of the light source (degrees). Defaults to
300.}

\item{normalize}{Logical. If `TRUE` (default), values below zero are set to
zero and the results are multiplied by 255.}

\item{overwrite}{Logical. If `TRUE` (and `filename` is set), overwrites
existing raster file.}

\item{filename}{Character. Optional output filename.}
}
\value{
A `RasterLayer`.
}
\description{
Can be used to calcualte a hillshade from a DEM `RasterLayer` of the type
returned by `get_dem()`.
}
