% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_is_type}
\alias{st_is_type}
\alias{st_is_multipolygon}
\alias{st_is_polygon}
\alias{st_is_multilinestring}
\alias{st_is_linestring}
\alias{sf_is_multipoint}
\alias{sf_is_point}
\title{Tests `sf` Dataframes for Geometry Types}
\usage{
st_is_type(df, type, exact = FALSE)

st_is_multipolygon(df, ...)

st_is_polygon(df, ...)

st_is_multilinestring(df, ...)

st_is_linestring(df, ...)

sf_is_multipoint(df, ...)

sf_is_point(df, ...)
}
\arguments{
\item{df}{An `sf` dataframe.}

\item{type}{Character. Type to match.}

\item{exact}{Boolean. If `FALSE` (default), returns unexact matches. (E.g., 
`"POLYGON"` will also match `"MULTIPOLYGON".`)}

\item{...}{Additional arguments passed to `st_is_type()`.}
}
\value{
Boolean.
}
\description{
`st_is_type()` returns `TRUE` if `type` matches geometry type returned by 
`sf::st_geometry_type()`.

`st_is_multipolygon()` returns `TRUE` if geometry column contains 
MULTIPOLYGONS.

`st_is_polygon()` returns `TRUE` if geometry column contains 
POLYGONS (or MULTIPOLYGONS, if `exact` is FALSE).

`st_is_multilinestring()` returns `TRUE` if geometry column contains 
MULTILINESTRINGS.

`st_is_linestring()` returns `TRUE` if geometry column contains 
LINESTRINGS (or MULTILINESTRINGS if `exact` is FALSE).

`st_is_multipoint()`  returns `TRUE` if geometry column contains 
MULTIPOINTS.

`st_is_point()` returns `TRUE` if geometry column contains 
POINTS (or MULTIPOINTS if `exact` is FALSE).
}
