% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_multi_type_center}
\alias{st_multi_type_center}
\title{Return Geometry Centers for Multiple Types.}
\usage{
st_multi_type_center(df, on_surface = TRUE, retain_geom = FALSE)
}
\arguments{
\item{df}{\code{sf} object}

\item{on_surface}{If \code{TRUE} (the default), generates polygon centroids using
\code{st_point_on_surface()}. If \code{FALSE}, uses \code{st_centroid()}.}

\item{retain_geom}{Boolean. If \code{TRUE} (the default), returns points. If
\code{FALSE} preserves geometry type of input.}
}
\value{
An \code{sf} dataframe.
}
\description{
Return Geometry Centers for Multiple Types.
}
