% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_preprocess}
\alias{st_preprocess}
\title{Standard preprocessing operations for spatial data.}
\usage{
st_preprocess(df, crs, name = "geometry")
}
\arguments{
\item{df}{`sf` object}

\item{crs}{EPSG code or `crs` object.}

\item{name}{Character. Name of geometry column. `"geometry"` is default.}
}
\value{
Simple Features dataframe.
}
\description{
Standard preprocessing operations for spatial data.
}
