% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_square_it}
\alias{st_square_it}
\title{Bounding Square, Bounding Circle}
\usage{
st_square_it(df, circle = FALSE)
}
\arguments{
\item{df}{`sf` object}

\item{circle}{If `TRUE`, returns a circle. If `FALSE` (the default), returns
a square.}
}
\value{
A circle or square `sf` object whose size is given by the larger
dimension of a layer's bounding box.
}
\description{
Bounding Square, Bounding Circle
}
