% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{st_xyxy_to_lines}
\alias{st_xyxy_to_lines}
\title{XY-XY To Lines}
\usage{
st_xyxy_to_lines(
  df,
  names = c("x_h", "y_h", "x_w", "y_w"),
  retain_cols = FALSE
)
}
\arguments{
\item{df}{A dataframe with four columns containing two XY pairs.}

\item{names}{list of the names of the four columns.}

\item{crs}{Coordinate reference system.}
}
\value{
A dataframe with LINESTRING geometries.
}
\description{
Takes a dataframe with four columns containing two XY pairs (e.g., [X_{1},
Y_{1}], [X_{2}, Y{2}]) and returns an sf dataframe with the same number of
rows with those coordinates converted into linestrings.
}
