% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigris.R
\name{tigris_get_*}
\alias{tigris_get_*}
\alias{tigris_get_states}
\alias{tigris_get_counties}
\alias{tigris_get_multistate}
\alias{tigris_get_multistate_by_county}
\alias{tigris_get_places}
\alias{tigris_get_tracts}
\alias{tigris_get_block_groups}
\alias{tigris_get_roads}
\alias{tigris_get_primary_roads}
\alias{tigris_get_primary_secondary_roads}
\alias{tigris_get_rails}
\alias{tigris_get_area_water}
\alias{tigris_get_linear_water}
\title{Get Census-Designated Places}
\usage{
tigris_get_states(states = NULL, crs = 4326, ...)

tigris_get_counties(states = NULL, crs = 4326, counties = NULL, ...)

tigris_get_multistate(.function, states, crs = 4326, ...)

tigris_get_multistate_by_county(
  .function,
  states,
  crs = 4326,
  counties = NULL,
  ...
)

tigris_get_counties(states = NULL, crs = 4326, counties = NULL, ...)

tigris_get_places(states, crs = 4326, ...)

tigris_get_tracts(states, crs = 4326, ...)

tigris_get_block_groups(states, crs = 4326, ...)

tigris_get_roads(states, crs = 4326, counties = NULL, ...)

tigris_get_primary_roads(crs = 4326, ...)

tigris_get_primary_secondary_roads(states, crs = 4326, ...)

tigris_get_rails(crs = 4326, ...)

tigris_get_area_water(states, crs = 4326, counties = NULL, ...)

tigris_get_linear_water(states, crs = 4326, counties = NULL, ...)
}
\arguments{
\item{states}{Vector of two-character state codes.}

\item{crs}{EPSG code or `crs` object. `4326` default.}

\item{...}{Passed on to `tigris` download function (e.g., 
`tigris::tracts()`).}

\item{counties}{Character vector of county names. If 
`NULL` (the default), all counties are returned.}

\item{.function}{`tigris` function to run.}
}
\value{
Simple Features dataframe.
}
\description{
`tigris_get_places()` retrieves census-designated places.

`tigris_get_states()` retrieves states, commonwealths, and territories.

`tigris_get_counties()` retrieves counties and equivalents.

`tigris_get_multistate()` is a helper function that downloads multiple states
worth of data, where `tigris` function prevents this.

`tigris_get_multistate_by_county()` is a helper function that downloads
multiple states worth of data, county-by-county, where `tigris` function 
prevents this.

`tigris_get_tracts()` retrieves census tracts.

`tigris_get_block_groups()` retrieves block_groups.

`tigris_get_roads()` retrieves roads.

`tigris_get_primary_roads()` retrieves primary roads.

`tigris_get_primary_secondary_roads()` retrieves primary/secondary roads.

`tigris_get_area_water()` retrieves areal water features.

`tigris_get_linear_water()` retrieves linear water features.
}
