% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigris.R
\name{tigris_get_}
\alias{tigris_get_}
\alias{tigris_get_states}
\alias{tigris_get_counties}
\alias{tigris_get_places}
\alias{tigris_get_postal}
\alias{tigris_get_tracts}
\alias{tigris_get_block_groups}
\alias{tigris_get_roads}
\alias{tigris_get_rails}
\alias{tigris_get_area_water}
\alias{tigris_get_linear_water}
\title{Get Census Geographies}
\usage{
tigris_get_states(places, year = NULL, crs = 4326, filter = TRUE)

tigris_get_counties(places, year = NULL, crs = 4326, filter = TRUE)

tigris_get_places(places, year = NULL, crs = 4326, filter = TRUE)

tigris_get_postal(places = NULL, year = 2010, crs = 4326)

tigris_get_tracts(places, year = NULL, crs = 4326)

tigris_get_block_groups(places, year = NULL, crs = 4326)

tigris_get_roads(places, year = NULL, crs = 4326)

tigris_get_rails(places, year = NULL, crs = 4326)

tigris_get_area_water(places, year = NULL, crs = 4326)

tigris_get_linear_water(places, year = NULL, crs = 4326)
}
\arguments{
\item{places}{A vector of cities, counties or states.}

\item{year}{Year of TIGER/Line geometries to pull.}

\item{crs}{EPSG code or \code{crs} object. \code{4326} default.}

\item{filter}{If \code{TRUE}, will filter result by query feature. (For example,
\code{tigris_get_states()} will return only requested states, even though
tigris::states() returns the entire nation). Note that this will not speed
the query, since the filtering happens after fetching the dataset.}
}
\value{
Simple Features dataframe.
}
\description{
\code{tigris_get_states()} retrieves states, commonwealths, and territories.

\code{tigris_get_counties()} retrieves counties and equivalents.

\code{tigris_get_places()} retrieves census-designated places.

\code{tigris_get_postal()} retrieves ZIP code tabulation areas.

\code{tigris_get_tracts()} retrieves census tracts.

\code{tigris_get_block_groups()} retrieves block groups.

\code{tigris_get_roads()} retrieves roads.

\code{tigris_get_rails()} retrieves  rail lines.

\code{tigris_get_area_water()} retrieves areal water features.

\code{tigris_get_linear_water()} retrieves linear water features.
}
